Add additional configuration options when creating index.html
We need to be able to set the base URL used by the Swagger UI when
making API calls, since it may vary at runtime. The swagger-ui repo
has an open issue on one way to accomplish this here:
https://github.com/swagger-api/swagger-ui/issues/5981

To implement this plugin and use it, we need to be able to adjust the
JavaScript being templated into index.html. In other to keep this
flexible, I've added new fields to the Config struct to optionally
add user specified JavaScript and/or configuration object properties.
These fields are all opt-in, so the existing behavior should not be
changed. The changes are also open ended enough that they could
be used for other requirements that may arise.